requirements.txt
ipykernel
openai
python-dotenv
streamlit
streamlit_webrtc
moviepy
pydub
pandas
matplotlib
Pillow
pyaudio
SpeechRecognition
sounddevice
playsound==1.2.2
